refactor(test_propagate_filter): restructure test_propagate_filter into several modules for splits
